+++ src/app/(routes)/ledgers/[id]/transactions/create/review/page.tsx
@@ -25,6 +25,7 @@ import Image from 'next/image'
 import { isNil } from 'lodash'
 import { useCreateTransaction } from '@/client/transactions'
 import { useOrganization } from '@/context/organization-provider/organization-provider-client'
+import useCustomToast from '@/hooks/use-custom-toast'
 import { TransactionFormSchema } from '../schemas'
 
 export default function CreateTransactionReviewPage() {
@@ -32,15 +33,28 @@ export default function CreateTransactionReviewPage() {
   const router = useRouter()
 
   const { id } = useParams<{ id: string }>()
+  const { showSuccess, showError } = useCustomToast()
 
   const { currentOrganization } = useOrganization()
 
-  const { mutate: createTransaction, isPending: createLoading } =
-    useCreateTransaction({
-      organizationId: currentOrganization.id!,
-      ledgerId: id,
-      onSuccess: () => router.push('/transactions')
-    })
+  const {
+    mutate: createTransaction,
+    isPending: createLoading,
+    error,
+    data
+  } = useCreateTransaction({
+    organizationId: currentOrganization.id!,
+    ledgerId: id,
+    onSuccess: (data) => {
+      showSuccess('Transaction created successfully')
+      handleSubmitClose()
+      router.push(`/ledgers/${data.ledgerId}/transactions/${data.id}`)
+    },
+    onError(message) {
+      showError(message)
+      handleSubmitClose()
+    }
+  })
 
   const { values, currentStep } = useTransactionForm()
 
@@ -62,10 +76,13 @@ export default function CreateTransactionReviewPage() {
     }))
   })
 
-  const { handleDialogOpen: handleSubmitOpen, dialogProps: submitDialogProps } =
-    useConfirmDialog({
-      onConfirm: () => createTransaction(parse(values))
-    })
+  const {
+    handleDialogOpen: handleSubmitOpen,
+    dialogProps: submitDialogProps,
+    handleDialogClose: handleSubmitClose
+  } = useConfirmDialog({
+    onConfirm: () => createTransaction(parse(values))
+  })
 
   return (
     <>