+++ src/schema/account.ts
@@ -11,7 +11,7 @@ const assetCode = z.string()
 
 const portfolioId = z.string().nullable().optional()
 
-const productId = z.string().nullable().optional()
+const segmentId = z.string().nullable().optional()
 
 const type = z.string()
 
@@ -25,7 +25,7 @@ export const accountSchema = z.object({
   entityId,
   assetCode,
   portfolioId,
-  productId,
+  segmentId,
   metadata,
   type,
   allowSending,